ndb - (refactor) remove unused member lhfragbits in dbacc
remove senseless use of lhfragbits

the use of lhfragbits to skip bits in hash value in dbacc
are removed. this was probably some legacy code. now
fragment are hashed from second word in the elements md5 sum.

more cleanup is possible, also in signals but i do not
do it in this patch.