ndb - introduce new integer type storing variable number of bits
it use an underlying integer type.
the most significant set bit marks the limit of valid bits,
only the lower significant bits are well defined.
The more significant bits will be unset, and they as well as
the top set bit will be treated as having unknown value.

the method match() returns true if the lower bits that is valid
for both arguments are equal, indicating that the full values
may be equal. Otherwise it returns false, meaning that the two
values definitly is unequal.

There are also methods for shifting out or in lower bits, and
to inspect one bit or take out some range of lower bits.