ndb - (refactor) change logic for expand/shrink control in dbacc
change expandFlag to expandShrinkQueued

The only values currently used for expandFlag was 0 and 2,
reenable_expand_after_redo_log_exection_complete() was never
called.

So i replaced the logic with on boolean flag expandShrinkQueued
that is set true on send of expand or shrink signal, and set
false when signal is processed. The flag is checked before send
of such signal and if set true, the send is suppressed, meaning
that at most one expand or shrink signal can be queued at the
time.