Bug#14471052 MYSQL_PLUGIN.EXE CANT FIND MY_PRINT_DEFAULTS.EXE IN CURRENT DIRECTORY
On Windows, MYSQL_PLUGIN.EXE is trying to find path of MY_PRINT_DEFAULTS.EXE 
using non-absoute path. It uses CRT function _stati64 for this purpose. However 
_stati64 works only with absolute path and as a result, MYSQL_PLUGIN.EXE cannot 
find the path of MY_PRINT_DEFAULTS.EXE.

I have now modified the code so that if the search using the relative path fails 
on Windows, we prepend the path of the executable and then try to search 
MY_PRINT_DEFAULTS.EXE in that path.

I have also added a test case. This test case is based on mysql_plugin.test.

This is a backport from the fix in mysql-trunk revid 
ravinder.thakur@oracle.com-20121001063602-csbaww1evsr07ive