Bug#14498914 CRASH IN ADD_FOUND_MATCH_TRIG_COND WITH _ROWID IN SUBQUERY
Problem: A query with a subquery in the JOIN ... ON clause with an
outer reference to a field that is out of scope may cause the server
to crash.

When an ON clause is encountered by the parser, a new
Name_resolution_context is created for this clause and pushed to the
context stack. This context limits resolving to the tables that are
visible in this ON clause. Field references in the ON clause will be
resolved using this context.

When the ON clause contains a subquery, mysql_new_select() creates a
new Name_resolution_context for the subquery, and this context is
linked with the context of the outer query. Later, when resolving
fields in the subquery, Item_field::fix_fields() will first look for
the field in the subquery context, then, if it is not found, it will
follow the context's outer_context pointer and try again with that
context. Since the outer context for the subquery is the outer query's
context, not the ON context, the resolver is able to find fields that
are out of scope in the ON clause and therefore also in the subquery.
Later on, the server crashes when it tries to follow a null pointer
caused by this invalid field reference.

Fix: Let subqueries in ON clauses have the ON context as outer
context.