Bug#14641759: Extra rows on second exec of prepared stmt with nested IN               (subquery) and view.
This is related to bug#13773979, which fixed a bug with prepared
statements that referenced views and performed semi-join transformation.
However, it only fixed bugs for which a view reference was semijoin-
transformed once. If transformation occurred more than once, the
select_lex stored in the Name_resolution_context would still be wrong.

Fixed by remembering which select_lex objects have been swallowed into
a parent select_lex and updating context->select_lex for each of them
to point to the parent select_lex.