Bug#14666396 EXCHANGE PARTITION W/TRIGGER CAUSING OOM
The problem was that issuing statements that caused a stored
program (such as a trigger) to be executed a large number of
times could cause the server to allocate excessive amounts of
memory.

The reason was that the stored program definition was reparsed
for each execution and each reparsing requires extra memory which
was not freed until the stored program was removed from memory.

Reparsing is supposed to happen when metadata of objects used
by a stored program instruction changes. For example by doing
ALTER TABLE on a table used by an instruction. This is
necessary to ensure correctness and was introduced by WL#4179.

The bug was that a flag indicating that the new metadata is
valid, was not correctly set after reparse for two types of
stored program instructions. This meant that once metadata
used by such instructions was invalidated once, it never
became valid again and therefore reparsed every time.

This patch fixes the problem by setting the proper flag after
reparse for sp_instr_stmt and sp_instr_copen similar to what
is already done for other types of stored program instructions.

No test case was added since the bug only manifests in increased
memory and CPU usage for which it is difficult to check with
MTR tests.