Worklog #6073: Preliminary patch
Fixes a bug in the (empty) initialization of the error message
array. The loop failed to set the last error message's string. This
will cause a core dump if we want to unit test code that raises the
last error (with code ER_ERROR_LAST). In unit testing we don't read
the error messages so we rely on the empty initialization.