Worklog #6073: Remove INSERT DELAYED: Fifth patch, fine-grained removal.
In this patch we remove what's left of the delayed insert
implementation in the server. The class Delayable_insert_operation is
removed, and the protected members that it changes are now made const
in COPY_INFO. The unit test for the class is also removed. In
sql_insert.cc we remove some leftover comments that still mention
delayed inserts, and the now-unused class delayed_row.