BUG#14600994: RANGE OPTIMIZER DOES NOT SET KEYPART_MAP               FOR GEO RANGES
sel_arg_range_seq_next() did not assign a value to the range's
keypart_map. Due to this, records_in_range() was called with a
keypart_map with random bits set. records_in_range() uses this
bitmap to build keys used as min and max value in the two index
dives, but since the bits are not set an ASSERT fires.

The fix is to set the keypart_map to the keypart bits used by
the range predicate.