Bug#14601361 READ AND COMPARE OF FREED MEMORY IN ITEM_FIELD::REGISTER_FIELD_IN_READ_MAP
Problem: A query with a WHERE clause with a subquery with an outer
reference as WHERE clause may cause the server to crash.

During resolving, JOIN::prepare() for the subquery calls
setup_without_group(), which again calls setup_conds() to resolve
items in the WHERE clause of the subquery. The WHERE clause consists
of a single Item_field that is resolved to a field in the view and
replaced by an Item_direct_view_ref. Since the Item_field is the only
item in the subquery's JOIN::conds, JOIN::conds is set to point to the
new Item_direct_view_ref. However, st_select_lex::where still points
to the Item_field.

When filesort()'s find_all_keys() later calls select->cond->walk(), it
calls Item_subselect::walk, which continues the walk on its
st_select_lex::where instead of JOIN::conds. The st_select_lex::where
still points to the Item_field, and the walk crashes when
Item_field::register_field_in_read_map dereferences a null pointer.

Fix: Set st_select_lex::where to point to the same item as JOIN::conds
after resolving in setup_conds().