Bug#14642248 NONSENSE ERROR ON WINDOWS (ERRNO: 22 - INVALID ARGUMENT)
On Windows, " is an invalid file name character. When a SQL command is given to
select a table with " character, MySQL tries to find a file name with " in it. 
For example, if the SQL select * from `t"`; is given, MySQL will try to find a file with 
name t".frm. Since t".frm is an invalid file name on Windows, Win32 API fails with error 
code ERROR_INVALID_NAME. However in my_winerr.c this error code is not recognized 
and hence MySQL thinks some unknown error has happened. As a result, MySQL logs 
the error message to server error log and throws back a generic error code/message
(ERRNO: 22 - INVALID ARGUMENT).

The fix is to recognize that if the failure was due to Win32 API returning 
ERROR_INVALID_NAME, it means that table does not exist. This is because:

a) Tables with invalid names cannot exist and
b) When MySQL tries to find table with name t".frm, it has already tried table 
search with normalized names(t@0202.frm) and that too has failed.

So we can surely say that if ERROR_INVALID_NAME is obtained during table search,
the table in question does not exist. We do this by mapping ERROR_INVALID_NAME
to ENOENT in my_winerr.c. This instructs MySQL to return error saying that this table 
does not exist. Also since now this is a known error, MySQL does not log it in error log.