bug#14207773 - MYSQLD --CONSOLE DON'T SHOW ANYTHINGS
In the 5.1 version of MySQL if both --console and --log-error options 
were given, messages from mysqld.exe were shown on the console. 
However after 5.3, if both options were specified, output did not go 
to console. This happened due to some code re factoring in revision 
2876.75.1.

opt_error_log variable controls whether the data sent to stderr should be 
redirected to the log file or not. After the refactoring this variable is
being set to 1 even if the the --console option is specified. As a result 
output to console is redirected to log file and nothing is seen on console.

Proposed fix sets opt_error_log flag only if the console option is not 
specified or mysqld.exe is running as a service. 

The changes also contains fix for some test cases that failed. The test 
cases failed because they depend on the creation of log file
(--log-error option). The log file was not getting created because MTR 
by default passes --console option to MTR. I have added the --no-console 
option for failing test cases so that the error log file gets created.

This is a windows only bug.