BUG#11755766 - MAIN.PERROR-WIN FAILS ON NON-ENGLISH VERSIONS OF                             WINDOWS
DESCRIPTION:
On Non-English systems perror-win test fails.


ANALYSIS :
Since this test generates Error Messages corresponding to
input Error codes based on the current System Language, when
it is run on Non-English systems the test fails because there
is a mismatch in the result generated.


FIX :
Code has been added to skip the test if run on a Non-English system.