Bug#14792821: CRASH IN WARNING_INFO::RESERVE_SPACE() WHEN RESIGNAL               WITH MAX_ERROR_COUNT=0
This crash occured if RESIGNAL with condition value was used when
max_error_count was 0. It happened because we try to make room
in the error list for the new condition RESIGNAL will raise.
However, we did not check if the error list actually had enough
conditions to remove. So we ended up trying to remove an element
from a list that was already empty.

This patch fixes the problem by checking if the error list
actually has any elements before trying to remove.