BUG#14666177 - TIME DIFF BETWEEN LAST 2 RECONNECTS IS                COMING > @@GLOBAL.NET_READ_TIMEOUT * 2
=== Problem ===

The test checks if the slave tries to reconnect before the
value of slave_net_timeout. The check was done using the
two variables:

1) ts_last: timestamp of last reconnect-attempt by slave,
2) ts_prev: timestamp of the second last reconnect-attempt
            by slave.

The slave threads were killed using
     --source include/stop_slave.inc
If the above statement is delayed due to CPU scheduling,
the slave would get enough time to try two reconnects.
If not, the slave threads would be killed after a single
reconection attempt by the slave.In this case,
     ts_last=ts_prev.

There was an assert condition given by:
     assert(@ts_last-@ts_prev >= slave_net_timeout
where slave_net_timeout=10.

In the above case,
     ts_last-ts_prev=0
which is less than slave_net_timeout. So, the assert condition
fired and the test failed.
 
=== Fix ===

1) Deleted the line
     --source include/stop_slave.inc
   which killed the slave threads earlier.  
2) The assert condition is replaced by a wait condition
   which makes sure that the slave makes two
   reconnection attempts.