Bug#14832472: Make mysqldump work with GTIDs
Problem
========
When a backup is taken using mysqldump on a server
with gtid enabled, the dump file does not contain
any gtid information.  This eventually results in 
replicating the transactions since history when the
backup is used to bring up a slave.
 
Solution
==========     
Fixed by adding SET @@GLOBAL.SET_GTID_PURGED=<value
of gtid_executed on the server> in the dump file.
For this, --set-gtid-purged has three possible values.
i) If --set-gtid-purged=OFF, output doesn't contain gtid info.
ii)If --set-gtid-purged=ON, and server has GTID_MODE!=ON, an
   error is generated.
iii)If --set-gtid-purged=AUTO, and server has GTID_MODE!=ON, this
   option does nothing. However, if GTIDs are enabled, the output
   contains GTID_PURGED info.