BUG#14711347 - TURNING SLOW QUERY LOG ON WITHOUT FILE PERMISSIONS                CRASHES SERVER
Description:

If SET GLOBAL slow_query_log='ON' command is given from MySQL
client to enable slow query logging, but the user does not have
write permission on the slow query logging file, MySQL server
crashes due to an assert failure. The same problem exists for
general_log option also.

Analysis:

When activate_log_handler() opens the file for logging, it is
not checking whether the file is successfully opened or not for
query logging.

Fix:

Added the condition on the open_slow_log() and open_query_log()
to check for errors in file opening. The return flag value is
changed based on finding of errors in opening value. Test case
added.