BUG#15912728: MYSQLBINLOG + -D + GTIDS => MAY OUTPUT NON-TERMINATED TRANSACTIONS
When mysqlbinlog is called with --database=X, there can be the
case that some (or even all) events for a given transaction are
suppressed from the output. Thence this filtering is dangerous,
even more with GTIDs, since it can suppress the transaction 
termination event: COMMIT.

This patch fixes mysqlbinlog to output a COMMIT event when
a GTID_LOG_EVENT is handled, provided that we have skipped
an event before and have not committed already. This works for
DML and DDL that commits implicitly. DDL that commits implicitly
is a single statment, so it is either suppressed completely or not.