Bug#15863023 SUPPORT IN-PLACE EXTENSION OF VARCHAR COLUMN (WL#6554)
The WL#5534 ALTER TABLE API introduces the flag
ALTER_COLUMN_EQUAL_PACK_LENGTH for extending a VARCHAR column so that the
character set and collation remains the same.

This flag is NOT set when the column is being extended from up to than 255
bytes to at least 256 bytes.
This is perfect for InnoDB, because when the maximum length is less than 256
bytes, the current length will be stored in one byte. When the 256-byte boundary
is not being crossed, we can extend the VARCHAR in-place, by just updating
the data dictionary.

rb:1547 approved by Jimmy Yang, Dmitry Lenev