Patch implementing for WL#6555 "Online rename index".
This patch adds support of RENAME INDEX operation to the ALTER TABLE
statement. It also extends the storage engine API for in-place ALTER
TABLE so storage engines can support this operation in-place
(i.e. without table rebuild). Code which determines if ALTER TABLE
can be done in-place for "simple" storage engines like MyISAM, Heap
and etc. was updated to handle ALTER TABLE ... RENAME INDEX as
an in-place operation. Support for in-place ALTER TABLE ... RENAME
INDEX for InnoDB tables will come as a separate patch. ALTER TABLE
... RENAME INDEX ... works for all storage engines when ALTER TABLE
is executed using COPY algorithm.

Syntax changes
==============

A new type of <alter_specification> is added:

<rename index clause> ::= RENAME ( INDEX | KEY ) <oldname> TO <newname>

Where <oldname> and <newname> are identifiers for old name and new 
name of the index.

Semantic changes
================

The result of "ALTER TABLE t1 RENAME INDEX a TO b" is a table which
contents and structure are identical to the old version of 't1' with
the only exception index 'a' being called 'b'.

Neither <oldname> nor <newname> can be "primary". The index being
renamed should exist and its new name should not be occupied
by another index on the same table.

Code/implementation changes
===========================

Added support of this clause in default implementation of ALTER
TABLE (i.e. using COPY algorithm) by changing mysql_prepare_alter_table()
code to adjust the name of the index being renamed when we construct the
key description for the new version of table. The process of copying
between old and new versions is not affected by this clause at all.

Support for this clause as in-place operation is added by:

* Introducing Alter_inplace_info::RENAME_INDEX flag for 
  Alter_inplace_info::handler_flags bitmap. This flag is set
  when at least one index is renamed by this ALTER TABLE. The
  flag is not set if index in question is changed otherwise.
* Adding Alter_inplace_info::index_rename_buffer/count members
  for array of of pairs of pointers to new and old versions of
  KEY for each index renamed (in
  Alter_inplace_info::key_buffer_info and TABLE::keyinfo
  correspondingly).
* Changing code in fill_alter_inplace_info() to set this flag
  and the array according to the ALTER TABLE statement.
* Changing stub implementation of
  handler::check_if_supported_inplace_alter() method to treat
  ALTER TABLE ... RENAME INDEX as an in-place operation, so
  "simple" storage engines that use this implementation (e.g.
  MyISAM, Heap, ...) can handle it as such.