BUG#15947962: DROP BOTH TEMPORARY TABLE AND REGULAR TABLES ON A SINGLE STATEMENT MISSES GTID
When temporary and regular tables or temporary tables with
different storage engines are dropped on a single
statement, the original statement is split in two.
These two statements are logged in two events to binary logs,
when gtid_mode is ON each DDL event must have its own GTID.

Since drop temporary table does not implicitly commit, in these
cases we must force a commit.