Bug#14475946: Cannot retrieve MySQL 5.0 stacktrace on Linux
In 5.5+, we use the backtrace() interface on linux
which was added to glibc in 2.1, circa 1999.
(At the time of this patch, we officially support glibc
2.3 and up.)

This package removes our homebrew solution, which is
now essentially dead code, and prints a message to the
DBA if they somehow manage to reach it anyway.

(Mind that this would presumably necessitate linking
against a stone age glibc, or a libc that is not GNU.
Either way, we fail gracefully in such cases.)