BUG#15992220: MAKE IT POSSIBLE TO DETERMINE IF GTID AUTO-POSITIONING PROTOCOL IS IN USE
When GTIDs were introduced, a new master-slave protocol with 
auto-positioning was also introduced. The new protocol is 
turned on by: "CHANGE MASTER TO MASTER_AUTO_POSITION = 1"

However, but there is currently no way to determine if the 
protocol is used or not.

This patch extends the fields in SHOW SLAVE STATUS with an
additional field: Auto_Position = (0|1), which states whether
the new protocol is in use (1) or not (0).